+++ worlds/pokemon_bw/generate/__init__.py
@@ -54,8 +54,13 @@ class SpeciesEntry:
     tm_hm_moves: TMHMMovesetData
     is_custom_form: bool
     custom_form_file: int
-    # bitflag, b0 = evolutions, b1 = plando evo override, b2 = base stats, b3 = catch rate, b4 = levelup moveset
     write: int = 0
+    """b0 = evolutions
+    b1 = plando evo override
+    b2 = base stats
+    b3 = catch rate
+    b4 = levelup moveset
+    b5 = types"""
     evo_line: EvoLine | None = None
 
     def __init__(self, name: str, data: SpeciesData):